+++ changelog.md
@@ -2,6 +2,7 @@ Version 0.4.2-beta
 - Fixed tooltip position not updating when anchor moved.
 - Tooltip is now hidden when scroll input is detected.
 - Added a multi-select tool.
+- Renamed SortableTable.RowPanel to SortableTable.TableRow
 
 Version 0.4.1-beta
 - Changed getEnabled to isEnabled inside UIClickable.